Move AnalysisFileFilter to filters
