+++ filters.go
@@ -2,6 +2,8 @@ package cleve
 
 import (
 	"fmt"
+	"path/filepath"
+	"regexp"
 	"time"
 )
 
@@ -131,6 +133,36 @@ func NewAnalysisFilter() AnalysisFilter {
 	return AnalysisFilter{
 		PaginationFilter: NewPaginationFilter(),
 	}
+}
+
+// Analysis file filtering
+type AnalysisFileFilter struct {
+	AnalysisId string           `bson:"analysis_id,omitzero" json:"analysis_id,omitzero"`
+	FileType   AnalysisFileType `bson:"type,omitzero" json:"type,omitzero"`
+	Level      AnalysisLevel    `bson:"level,omitzero" json:"level,omitzero"`
+	ParentId   string           `bson:"parent_id,omitzero" json:"parent_id,omitzero"`
+	Name       string           `bson:"name,omitzero" json:"name,omitzero"`
+	Pattern    *regexp.Regexp   `bson:"-" json:"-"`
+}
+
+func (f *AnalysisFileFilter) Apply(file AnalysisFile) bool {
+	pass := true
+	if f.FileType.IsValid() && f.FileType != file.FileType {
+		return false
+	}
+	if f.Level.IsValid() && f.Level != file.Level {
+		return false
+	}
+	if f.ParentId != "" && f.ParentId != file.ParentId {
+		return false
+	}
+	if f.Name != "" && f.Name != filepath.Base(file.Path) {
+		return false
+	}
+	if f.Pattern != nil {
+		pass = f.Pattern.Match([]byte(file.Path))
+	}
+	return pass
 }
 
 // Sample filtering.
